Metabolites paste sets correct values when paste...
area contains a sorted column. It may be confusing the way the values
are sorted. Will put in tutorial a warning about pasting into areas with
sorted columns that the results may not be what is expected. If we get
complaints about this, there are other possibilities such as copying the
whole table and setting the new copy unsorted.